move hooks & update paths
